+++ frontend/src/components/Chart/ClusteringCard/StatsPanel/StatsPanel.tsx
@@ -1,7 +1,7 @@
 import * as S from './StatsPanel.style';
 import StatsItem from './StatsItem';
 import { useEffect, useState } from 'react';
-import { statsItem } from '@/types/chartType';
+import { statsItem } from '@/types/chart';
 import { STATSLABEL } from '@/constants/labelList';
 interface StatsPanelProps {
   isVisible: boolean;
